ft: add export CSV button to types table
