+++ apps/hash-frontend/src/shared/table-header.tsx
@@ -1,3 +1,4 @@
+import { SizedGridColumn } from "@glideapps/glide-data-grid";
 import {
   CheckIcon,
   Chip,
@@ -12,6 +13,7 @@ import {
   Entity,
   EntityTypeWithMetadata,
   extractOwnedByIdFromEntityId,
+  isBaseUrl,
   isExternalOntologyElementMetadata,
   PropertyTypeWithMetadata,
 } from "@local/hash-subgraph";
@@ -31,12 +33,18 @@ import {
 import {
   Dispatch,
   FunctionComponent,
+  MutableRefObject,
   ReactNode,
   SetStateAction,
+  useCallback,
   useMemo,
   useState,
 } from "react";
 
+import { Row } from "../components/grid/utils/rows";
+import { MinimalUser } from "../lib/user-and-org";
+import { TypeEntitiesRow } from "../pages/shared/entities-table/use-entities-table";
+import { TypesTableRow } from "../pages/types/[[...type-kind]].page/types-table";
 import { EarthAmericasRegularIcon } from "./icons/earth-americas-regular";
 import { FilterListIcon } from "./icons/filter-list-icon";
 import { HouseRegularIcon } from "./icons/house-regular-icon";
@@ -105,6 +113,11 @@ export type FilterState = {
   includeGlobal: boolean;
 };
 
+export type GetAdditionalCsvDataFunction = () => Promise<{
+  prependedData: string[][];
+  appendedData: string[][];
+} | null>;
+
 type TableHeaderProps = {
   internalWebIds: (AccountId | AccountGroupId)[];
   itemLabelPlural: "entities" | "pages" | "types";
@@ -122,7 +135,10 @@ type TableHeaderProps = {
   )[];
   filterState: FilterState;
   endAdornment?: ReactNode;
-  generateCsvFile?: GenerateCsvFileFunction;
+  title: string;
+  columns: SizedGridColumn[];
+  currentlyDisplayedRowsRef: MutableRefObject<Row[] | null>;
+  getAdditionalCsvData?: GetAdditionalCsvDataFunction;
   setFilterState: Dispatch<SetStateAction<FilterState>>;
   toggleSearch?: () => void;
   onBulkActionCompleted?: () => void;
@@ -141,7 +157,10 @@ export const TableHeader: FunctionComponent<TableHeaderProps> = ({
   selectedItems,
   filterState,
   endAdornment,
-  generateCsvFile,
+  title,
+  columns,
+  currentlyDisplayedRowsRef,
+  getAdditionalCsvData,
   setFilterState,
   toggleSearch,
   onBulkActionCompleted,
@@ -168,6 +187,81 @@ export const TableHeader: FunctionComponent<TableHeaderProps> = ({
     items && typeof numberOfUserWebItems !== "undefined"
       ? items.length - numberOfUserWebItems
       : undefined;
+
+  const generateCsvFile = useCallback<GenerateCsvFileFunction>(async () => {
+    const currentlyDisplayedRows = currentlyDisplayedRowsRef.current;
+    if (!currentlyDisplayedRows) {
+      return null;
+    }
+
+    const additionalCsvData = getAdditionalCsvData
+      ? await getAdditionalCsvData()
+      : undefined;
+
+    const { prependedData, appendedData } = additionalCsvData ?? {};
+
+    const prependedColumnTitles = prependedData?.[0];
+    const prependedRows = prependedData?.splice(1);
+
+    const appendedColumnTitles = appendedData?.[0];
+    const appendedRows = appendedData?.splice(1);
+
+    // Entity metadata columns (i.e. what's already being displayed in the entities table)
+
+    const columnRowKeys = columns.map(({ id }) => id).flat();
+
+    const tableContentColumnTitles = columns.map((column) =>
+      /**
+       * If the column is the entity label column, add the word "label" to the
+       * column title. Otherwise we'd end up with an "Entity" or "Page" column title,
+       * making it harder to distinguish from the property/outgoing link columns.
+       */
+      column.id === "entityLabel" ? `${column.title} label` : column.title,
+    );
+
+    // Collate the contents of the CSV file row by row (including the header)
+    const content: string[][] = [
+      [
+        ...(prependedColumnTitles ?? []),
+        ...tableContentColumnTitles,
+        ...(appendedColumnTitles ?? []),
+      ],
+      ...currentlyDisplayedRows.map((row, i) => {
+        const tableCells = columnRowKeys.map((key) => {
+          const value = row[key];
+
+          if (typeof value === "string") {
+            return value;
+          } else if (key === "lastEditedBy") {
+            const user = value as MinimalUser | undefined;
+
+            return user?.preferredName ?? "";
+          } else if (isBaseUrl(key)) {
+            /**
+             * If the key is a base URL, then the value needs to be obtained
+             * from the nested `properties` field on the row.
+             */
+            return (row as TypeEntitiesRow).properties?.[key] ?? "";
+          } else if (key === "archived") {
+            return (row as TypesTableRow).archived ? "Yes" : "No";
+          }
+
+          return "";
+        });
+
+        const prependedCells = prependedRows?.[i];
+        const appendedCells = appendedRows?.[i];
+
+        return [
+          ...(prependedCells ?? []),
+          ...tableCells,
+          ...(appendedCells ?? []),
+        ];
+      }),
+    ];
+
+    return { title, content };
+  }, [title, columns, currentlyDisplayedRowsRef, getAdditionalCsvData]);
 
   return (
     <Box
@@ -334,9 +428,7 @@ export const TableHeader: FunctionComponent<TableHeaderProps> = ({
             </Box>
           </Box>
         </Box>
-        {generateCsvFile ? (
-          <ExportToCsvButton generateCsvFile={generateCsvFile} />
-        ) : null}
+        <ExportToCsvButton generateCsvFile={generateCsvFile} />
         {endAdornment}
       </Box>
     </Box>
